requirement.txt
accelerate==0.24.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
asyncio==3.4.3
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
diffusers==0.11.1
dnspython==2.4.2
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.104.1
filelock==3.13.1
fsspec==2023.10.0
ftfy==6.1.3
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.2
huggingface-hub==0.19.4
idna==3.6
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
orjson==3.9.10
packaging==23.2
Pillow==10.1.0
psutil==5.9.6
pydantic==2.5.2
pydantic-extra-types==2.1.0
pydantic-settings==2.1.0
pydantic_core==2.14.5
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
redis==5.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.1
scipy==1.11.4
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
tokenizers==0.15.0
torch==2.1.1
torchvision==0.16.1
tqdm==4.66.1
transformers==4.35.2
triton==2.1.0
typing_extensions==4.8.0
ujson==5.8.0
urllib3==2.1.0
uuid==1.30
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
wcwidth==0.2.12
websockets==12.0
zipp==3.17.0
